Initial separation of concerns.
